+++ src/main/java/net/forscherfreunde/mod/entity/custom/CloudyEntity.java
@@ -1,12 +1,15 @@
 package net.forscherfreunde.mod.entity.custom;
 
 import net.forscherfreunde.mod.entity.vorlagen.CustomMobVorlage;
+import net.forscherfreunde.mod.entity.vorlagen.ModMob;
 import net.minecraft.entity.*;
 import net.minecraft.entity.ai.goal.*;
+import net.minecraft.entity.attribute.DefaultAttributeContainer;
 import net.minecraft.entity.attribute.EntityAttributeModifier;
 import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.damage.DamageSource;
 import net.minecraft.entity.mob.HostileEntity;
+import net.minecraft.entity.mob.MobEntity;
 import net.minecraft.entity.mob.ZombieEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.server.world.ServerWorld;
@@ -17,9 +20,17 @@ import net.minecraft.world.GameRules;
 import net.minecraft.world.SpawnHelper;
 import net.minecraft.world.World;
 
-public class CloudyEntity extends CustomMobVorlage {
+public class CloudyEntity extends ModMob {
     public CloudyEntity(EntityType<? extends HostileEntity> entityType, World world) {
         super(entityType, world);
+    }
+
+    public static DefaultAttributeContainer.Builder createMobAttributes() {
+        //hier werden die Basisattribute des Mobs gesetzt - hier könnt ihr diese also ganz einfach anpassen.
+        return MobEntity.createMobAttributes()
+                .add(EntityAttributes.GENERIC_MOVEMENT_SPEED, 0.7d)
+                .add(EntityAttributes.GENERIC_ATTACK_DAMAGE, 5)
+                .add(EntityAttributes.GENERIC_ATTACK_KNOCKBACK, 1);
     }
 
     @Override
@@ -34,57 +45,6 @@ public class CloudyEntity extends CustomMobVorlage {
             this.experiencePoints = (int)((double)this.experiencePoints * 2.5);
         }
         return super.getXpToDrop();
-    }
-
-    @Override
-    public boolean damage(DamageSource source, float amount) {
-        if (!super.damage(source, amount)) {
-            return false;
-        }
-        if (!(this.getWorld() instanceof ServerWorld)) {
-            return false;
-        }
-        ServerWorld serverWorld = (ServerWorld)this.getWorld();
-        LivingEntity livingEntity = this.getTarget();
-        if (livingEntity == null && source.getAttacker() instanceof LivingEntity) {
-            livingEntity = (LivingEntity)source.getAttacker();
-        }
-        if (livingEntity != null && this.getWorld().getDifficulty() == Difficulty.HARD && (double)this.random.nextFloat() < this.getAttributeValue(EntityAttributes.ZOMBIE_SPAWN_REINFORCEMENTS) && this.getWorld().getGameRules().getBoolean(GameRules.DO_MOB_SPAWNING)) {
-            int i = MathHelper.floor(this.getX());
-            int j = MathHelper.floor(this.getY());
-            int k = MathHelper.floor(this.getZ());
-            ZombieEntity zombieEntity = new ZombieEntity(this.getWorld());
-            for (int l = 0; l < 50; ++l) {
-                int m = i + MathHelper.nextInt(this.random, 7, 40) * MathHelper.nextInt(this.random, -1, 1);
-                int n = j + MathHelper.nextInt(this.random, 7, 40) * MathHelper.nextInt(this.random, -1, 1);
-                int o = k + MathHelper.nextInt(this.random, 7, 40) * MathHelper.nextInt(this.random, -1, 1);
-                BlockPos blockPos = new BlockPos(m, n, o);
-                EntityType<?> entityType = zombieEntity.getType();
-                SpawnRestriction.Location location = SpawnRestriction.getLocation(entityType);
-                if (!SpawnHelper.canSpawn(location, this.getWorld(), blockPos, entityType) || !SpawnRestriction.canSpawn(entityType, serverWorld, SpawnReason.REINFORCEMENT, blockPos, this.getWorld().random)) continue;
-                zombieEntity.setPosition(m, n, o);
-                if (this.getWorld().isPlayerInRange(m, n, o, 7.0) || !this.getWorld().doesNotIntersectEntities(zombieEntity) || !this.getWorld().isSpaceEmpty(zombieEntity) || this.getWorld().containsFluid(zombieEntity.getBoundingBox())) continue;
-                zombieEntity.setTarget(livingEntity);
-                zombieEntity.initialize(serverWorld, this.getWorld().getLocalDifficulty(zombieEntity.getBlockPos()), SpawnReason.REINFORCEMENT, null, null);
-                serverWorld.spawnEntityAndPassengers(zombieEntity);
-                this.getAttributeInstance(EntityAttributes.ZOMBIE_SPAWN_REINFORCEMENTS).addPersistentModifier(new EntityAttributeModifier("Zombie reinforcement caller charge", -0.05f, EntityAttributeModifier.Operation.ADDITION));
-                zombieEntity.getAttributeInstance(EntityAttributes.ZOMBIE_SPAWN_REINFORCEMENTS).addPersistentModifier(new EntityAttributeModifier("Zombie reinforcement callee charge", -0.05f, EntityAttributeModifier.Operation.ADDITION));
-                break;
-            }
-        }
-        return true;
-    }
-
-    @Override
-    public boolean tryAttack(Entity target) {
-        boolean bl = super.tryAttack(target);
-        if (bl) {
-            float f = this.getWorld().getLocalDifficulty(this.getBlockPos()).getLocalDifficulty();
-            if (this.getMainHandStack().isEmpty() && this.isOnFire() && this.random.nextFloat() < f * 0.3f) {
-                target.setOnFireFor(2 * (int)f);
-            }
-        }
-        return bl;
     }
 
 }
